requirements.txt
asgiref==3.7.2
Django==4.0.10
djangorestframework==3.13.0
install==1.3.5
psycopg2-binary==2.9.3
pytz==2024.1
sqlparse==0.4.4
django-cors-headers==3.10.0
dj-rest-auth==2.1.11
django-allauth==0.48.0
drf-spectacular==0.21.0
openai==1.13.3
